+++ 5/src/projection/mercator.h
@@ -16,48 +16,48 @@ namespace projection
 namespace mercator
 {
 
-static const float EQUATORIAL_RADIUS    = 6378137.0f;
-static const float POLAR_RADIUS         = 6356752.3142f;
-static const float RADIUS_RATIO         = POLAR_RADIUS / EQUATORIAL_RADIUS;
-static const float ECCENT               = sqrtf(1.0f - RADIUS_RATIO * RADIUS_RATIO);
-static const float COM                  = 0.5f * ECCENT;
+static const double EQUATORIAL_RADIUS    = 6378137.0;
+static const double POLAR_RADIUS         = 6356752.3142;
+static const double RADIUS_RATIO         = POLAR_RADIUS / EQUATORIAL_RADIUS;
+static const double ECCENT               = sqrt(1.0 - RADIUS_RATIO * RADIUS_RATIO);
+static const double COM                  = 0.5 * ECCENT;
 
 inline
-float lonToMet(float lon)
+double lonToMet(double lon)
 {
-    return EQUATORIAL_RADIUS * lon * M_PI / 180.0f;
+    return EQUATORIAL_RADIUS * lon * M_PI / 180.0;
 }
 
 inline
-float latToMet(float lat)
+double latToMet(double lat)
 {
-    lat = fmin(89.5f, fmax(lat, -89.5f));
-    float phi       = lat * M_PI / 180.0f;
-    float con       = ECCENT * sinf(phi);
-    con = powf((1.0f - con) / (1.0 + con), COM);
-    return 0.0f - EQUATORIAL_RADIUS * logf(tanf(0.5f * (M_PI * 0.5f - phi)) / con);
+    lat = fmin(89.5, fmax(lat, -89.5));
+    double phi       = lat * M_PI / 180.0;
+    double con       = ECCENT * sin(phi);
+    con = pow((1.0 - con) / (1.0 + con), COM);
+    return 0.0 - EQUATORIAL_RADIUS * log(tan(0.5 * (M_PI * 0.5 - phi)) / con);
 }
 
 inline
-float metToLon(float x)
+double metToLon(double x)
 {
-    return 180.0f * x / M_PI;
+    return 180.0 * x / M_PI;
 }
 
 inline
-float metToLat(float y)
+double metToLat(double y)
 {
-    float ts    = expf(-y / POLAR_RADIUS);
-    float phi   = M_PI_2 - 2.0f * atanf(ts);
-    float con   = ECCENT * sinf(phi);
-    float dphi  = 1.0f;
-    for(int i = 0; fabs(dphi) > 0.000000001f && i < 15; ++ i)
+    double ts    = expf(-y / POLAR_RADIUS);
+    double phi   = M_PI_2 - 2.0 * atanf(ts);
+    double con   = ECCENT * sinf(phi);
+    double dphi  = 1.0;
+    for(int i = 0; fabs(dphi) > 0.000000001 && i < 15; ++ i)
     {
-        dphi = M_PI_2 - 2.0f * atanf(ts * powf((1.0f - con) / (1.0f + con), COM)) - phi;
+        dphi = M_PI_2 - 2.0 * atan(ts * pow((1.0 - con) / (1.0 + con), COM)) - phi;
         phi += dphi;
     }
 
-    return 180.0f * phi / M_PI;
+    return 180.0 * phi / M_PI;
 }
 
 } // namespace mercator
